requirements.txt
asgiref==3.3.4
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
Django==3.2.3
django-tinymce==3.3.0
factory-boy==3.2.0
Faker==8.4.0
iniconfig==1.1.1
isort==5.8.0
mixer==7.1.2
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.3.0
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
